+++ Bit6.framework/Versions/A/Headers/Bit6FileDownloader.h
@@ -16,13 +16,14 @@ NS_ASSUME_NONNULL_BEGIN
 /*! Cancels all pending download operations. */
 + (void)cancelOperations;
 
-/*! Download a file. It can be set to replace any file in the indicated path if necessary.
+/*! Download a file. It can be set to replace any file in the indicated path if necessary. When running from an extension, a copy of the file will be saved in the SDK group container.
  @param url remote URL address of the file to download.
  @param filePath local path where to save the file.
  @param replace the file will be downloaded and will replace the file in cache.
  @param preparationBlock block to process the downloaded data before saving it to the filePath.
  @param priority the execution priority of the operation in the operation queue.
  @param completionHandler Block to call after the operation has been completed. The "error" value can be use to know if the file was downloaded and saved in cache.
+ @see +[Bit6 setAppGroupIdentifier:].
  */
 + (void)downloadFileAtURL:(NSURL*)url toFilePath:(NSString*)filePath canReplace:(BOOL)replace preparationBlock:(nullable NSData* (^)(NSData* data))preparationBlock priority:(NSOperationQueuePriority)priority completionHandler:(nullable void(^)(NSURL * _Nullable location, NSURLResponse * _Nullable response, NSError * _Nullable error))completionHandler;