+++ UISamples/Bit6UIFullDemo-ObjC/ChatDemo/InCall/CustomCallViewController.m
@@ -37,7 +37,7 @@
 
 - (Bit6CallController*) callController
 {
-    return [[Bit6 activeCalls] firstObject];
+    return [self.callControllers firstObject];
 }
 
 - (void)viewDidLoad {
@@ -45,7 +45,7 @@
     
     [self refreshState];
     
-    if ([Bit6 activeCalls].count>1) {
+    if (self.callControllers.count>1) {
         self.usernameLabel.address = nil;
         self.usernameLabel.text = @"Many Destinations";
     }
@@ -72,7 +72,7 @@
     [self.view addGestureRecognizer:tgr2];
     
     BOOL atLeastOneCallConnected = NO;
-    NSArray<Bit6CallController*>* callControllers = [Bit6 activeCalls];
+    NSArray<Bit6CallController*>* callControllers = self.callControllers;
     for (Bit6CallController *call in callControllers) {
         if (call.state >= Bit6CallState_CONNECTED) {
             atLeastOneCallConnected = YES; break;
@@ -85,7 +85,7 @@
 {
     BOOL atLeastOneCallConnected = NO;
     BOOL atLeastOneCallHasVideo = NO;
-    NSArray<Bit6CallController*>* callControllers = [Bit6 activeCalls];
+    NSArray<Bit6CallController*>* callControllers = self.callControllers;
     for (Bit6CallController *call in callControllers) {
         if (call.state >= Bit6CallState_CONNECTED) {
             atLeastOneCallConnected = YES;
@@ -127,7 +127,7 @@
     [super secondsDidChangeForCallController:callController];
     
     CGFloat longerCall = 0.0f;
-    NSArray<Bit6CallController*>* callControllers = [Bit6 activeCalls];
+    NSArray<Bit6CallController*>* callControllers = self.callControllers;
     for (Bit6CallController *call in callControllers) {
         if (call.seconds > longerCall) {
             longerCall = call.seconds;
@@ -175,7 +175,7 @@
 {
     Bit6CallController *smallerCall = nil;
     Bit6CallState smallerState = Bit6CallState_END;
-    NSArray<Bit6CallController*>* callControllers = [Bit6 activeCalls];
+    NSArray<Bit6CallController*>* callControllers = self.callControllers;
     for (Bit6CallController *call in callControllers) {
         if (call.state < smallerState) {
             smallerState = call.state;
@@ -231,7 +231,7 @@
     BOOL atLeastOneCallHasVideo = NO;
     BOOL atLeastOneCallHasAudio = NO;
     BOOL atLeastOneCallHasRemoteAudio = NO;
-    NSArray<Bit6CallController*>* callControllers = [Bit6 activeCalls];
+    NSArray<Bit6CallController*>* callControllers = self.callControllers;
     for (Bit6CallController *call in callControllers) {
         if (call.hasVideo) {
             atLeastOneCallHasVideo = YES;
@@ -287,7 +287,7 @@
 {
     [super updateLayoutForVideoFeedViews:videoFeedViews];
     
-    if ([Bit6 activeCalls].count>1) {
+    if (self.callControllers.count>1) {
         self.usernameLabel.address = nil;
         self.usernameLabel.text = @"Many Destinations";
     }