+++ UISamples/Bit6UIFullDemo-Swift/Bit6ChatDemo-Swift/CustomCallViewController.swift
@@ -33,7 +33,7 @@ class CustomCallViewController: Bit6CallViewController {
     
     var callController : Bit6CallController? {
         get {
-            return Bit6.activeCalls().first
+            return self.callControllers.first
         }
     }
     
@@ -42,7 +42,7 @@ class CustomCallViewController: Bit6CallViewController {
         
         refreshState()
         
-        if Bit6.activeCalls().count>1 {
+        if self.callControllers.count>1 {
             self.usernameLabel.address = nil
             self.usernameLabel.text = "Many Destinations"
         }
@@ -70,7 +70,7 @@ class CustomCallViewController: Bit6CallViewController {
         view.addGestureRecognizer(tgr2)
         
         var atLeastOneCallConnected = false
-        let callControllers = Bit6.activeCalls()
+        let callControllers = self.callControllers
         for call in callControllers {
             if call.state.rawValue >= Bit6CallState.CONNECTED.rawValue {
                 atLeastOneCallConnected = true
@@ -85,7 +85,7 @@ class CustomCallViewController: Bit6CallViewController {
         var atLeastOneCallConnected = false
         var atLeastOneCallHasVideo = false
         
-        let callControllers = Bit6.activeCalls()
+        let callControllers = self.callControllers
         for call in callControllers {
             if call.state.rawValue >= Bit6CallState.CONNECTED.rawValue {
                 atLeastOneCallConnected = true
@@ -128,7 +128,7 @@ class CustomCallViewController: Bit6CallViewController {
         
         if isViewLoaded() {
             var longerCall : UInt = 0
-            let callControllers = Bit6.activeCalls()
+            let callControllers = self.callControllers
             for call in callControllers {
                 if call.seconds > longerCall {
                     longerCall = call.seconds
@@ -177,7 +177,7 @@ class CustomCallViewController: Bit6CallViewController {
         var smallerCall : Bit6CallController? = nil
         var smallerState = Bit6CallState.END
         
-        let callControllers = Bit6.activeCalls()
+        let callControllers = self.callControllers
         for call in callControllers {
             if call.state.rawValue < smallerState.rawValue {
                 smallerState = call.state
@@ -235,7 +235,7 @@ class CustomCallViewController: Bit6CallViewController {
         var atLeastOneCallHasAudio = false
         var atLeastOneCallHasRemoteAudio = false
         
-        let callControllers = Bit6.activeCalls()
+        let callControllers = self.callControllers
         for call in callControllers {
             if call.hasVideo {
                 atLeastOneCallHasVideo = true
@@ -290,7 +290,7 @@ class CustomCallViewController: Bit6CallViewController {
     override func updateLayoutForVideoFeedViews(videoFeedViews: [Bit6VideoFeedView]) {
         super.updateLayoutForVideoFeedViews(videoFeedViews)
         
-        if Bit6.activeCalls().count>1 {
+        if self.callControllers.count>1 {
             self.usernameLabel.address = nil
             self.usernameLabel.text = "Many Destinations"
         }